Use more generic selectors to log in
Mizuho offers a white-labeled version of Money Forward which has a different log in page. (https://mizuho.x.moneyforward.com/users/sign_in)

Using type email and password allows me to keep a single version of the code that works with different log in pages. (Note however that this commit does not make the URL configurable, which would be necessary to fetch transactions from Money Forward for Mizuho.)

The downside is that somehow the page does not wait for log in to complete anymore, so I added a retry/timeout system that checks on the presence of a logout link in the page source.